+++ lib/src/ui/pluto_base_column_group.dart
@@ -108,6 +108,7 @@ class _ColumnGroupTitle extends StatelessWidget {
       width: groupTitleWidth,
       padding: EdgeInsets.symmetric(horizontal: _padding),
       decoration: BoxDecoration(
+        color: columnGroup.group.backgroundColor,
         border: Border(
           right: stateManager.configuration!.enableColumnBorder
               ? BorderSide(